requirements.txt
discord.py==2.3.2
twilio==8.10.0
google-cloud-speech==2.24.0
google-cloud-texttospeech==2.16.0
websockets==12.0
opuslib==3.0.1
pydub==0.25.1
pvporcupine==3.0.0
aiosqlite==0.19.0
cryptography==41.0.7
python-dotenv==1.0.0
numpy==1.26.2
scipy==1.11.4
